backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
pillow
python-dotenv
python-multipart
